Adiciona casts para silenciar avisos do MSVC 64-bit

+++ source/drawtext.c
@@ -60,7 +60,7 @@ HB_FUNC(TEXTOUT)
   void *hText;
   HB_SIZE nLen;
   LPCTSTR lpText = HB_PARSTR(4, &hText, &nLen);
-  TextOut(hwg_par_HDC(1), hwg_par_int(2), hwg_par_int(3), lpText, nLen); // TODO: o retorno é BOOL
+  TextOut(hwg_par_HDC(1), hwg_par_int(2), hwg_par_int(3), lpText, (int)nLen); // TODO: o retorno é BOOL
   hb_strfree(hText);
 }
 
@@ -90,7 +90,7 @@ HB_FUNC(DRAWTEXT)
     Array2Rect(hb_param(3, HB_IT_ARRAY), &rc);
   }
 
-  heigh = DrawText(hwg_par_HDC(1), lpText, nLen, &rc, uFormat);
+  heigh = DrawText(hwg_par_HDC(1), lpText, (int)nLen, &rc, uFormat);
   hb_strfree(hText);
 
   // if (HB_ISBYREF(uiPos))
@@ -162,7 +162,7 @@ HB_FUNC(GETTEXTSIZE)
   PHB_ITEM aMetr = hb_itemArrayNew(2);
   PHB_ITEM temp;
 
-  GetTextExtentPoint32(hwg_par_HDC(1), lpText, nLen, &sz); // TODO: o retorno é BOOL
+  GetTextExtentPoint32(hwg_par_HDC(1), lpText, (int)nLen, &sz); // TODO: o retorno é BOOL
   hb_strfree(hText);
 
   temp = hb_itemPutNL(NULL, sz.cx);
@@ -346,7 +346,7 @@ HB_FUNC(EXTTEXTOUT)
   rc.top = hb_parni(5);
   rc.right = hb_parni(6);
   rc.bottom = hb_parni(7);
-  ExtTextOut(hwg_par_HDC(1), hwg_par_int(2), hwg_par_int(3), ETO_OPAQUE, &rc, lpText, nLen, NULL);
+  ExtTextOut(hwg_par_HDC(1), hwg_par_int(2), hwg_par_int(3), ETO_OPAQUE, &rc, lpText, (UINT)nLen, NULL);
   hb_strfree(hText);
 }
